Update test cases for non parametric controlled ops
