+++ pennylane/ops/op_math/controlled.py
@@ -24,7 +24,6 @@ import numpy as np
 from scipy import sparse
 
 import pennylane as qml
-import pennylane.ops.op_math.controlled_non_parametric_ops
 from pennylane import operation
 from pennylane import math as qmlmath
 from pennylane.operation import Operator
@@ -129,8 +128,8 @@ def ctrl(op, control, control_values=None, work_wires=None):
         return ops_loader.ctrl(op, control, control_values=control_values, work_wires=work_wires)
 
     custom_controlled_ops = {
-        (qml.PauliZ, 1): pennylane.ops.op_math.controlled_non_parametric_ops.CZ,
-        (qml.PauliY, 1): pennylane.ops.op_math.controlled_non_parametric_ops.CY,
+        (qml.PauliZ, 1): qml.CZ,
+        (qml.PauliY, 1): qml.CY,
         (qml.PauliX, 1): qml.CNOT,
         (qml.PauliX, 2): qml.Toffoli,
         (qml.RX, 1): qml.CRX,
